requirements.txt
# MCP SDK
mcp>=0.9.0

# Google OAuth and API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0

# Environment variables
python-dotenv>=1.0.0
